requirements.txt

python-telegram-bot==13.15
flask==2.2.5
openai==1.3.5
